+++ frontend/src/app/_services/application.service.ts
@@ -9,12 +9,13 @@ import { environment } from '../../environments/environment';
 import { Router } from '@angular/router';
 
 import { SpecialUseApplication } from '../_models/special-use-application';
+import { UtilService } from './util.service';
 
 @Injectable()
 export class ApplicationService {
   private endpoint = environment.apiUrl + 'permits/applications';
 
-  constructor(private http: Http, public router: Router) {}
+  constructor(private http: Http, public router: Router, public util: UtilService) {}
 
   create(body, type, multipart = false) {
     let headers = new Headers({ 'Content-Type': 'application/json' });
@@ -26,21 +27,21 @@ export class ApplicationService {
     return this.http
       .post(this.endpoint + type, body, options)
       .map((res: Response) => res.json())
-      .catch(this.handleError);
+      .catch(this.util.handleError);
   }
 
   get(params = '') {
     return this.http
       .get(this.endpoint + params, { withCredentials: true })
       .map((res: Response) => res.json())
-      .catch(this.handleError);
+      .catch(this.util.handleError);
   }
 
   getOne(id, params = '') {
     return this.http
       .get(this.endpoint + params + id, { withCredentials: true })
       .map((res: Response) => res.json())
-      .catch(this.handleError);
+      .catch(this.util.handleError);
   }
 
   update(body, type) {
@@ -51,7 +52,7 @@ export class ApplicationService {
     return this.http
       .put(`${this.endpoint}/special-uses/${type}/${body.appControlNumber}`, body, options)
       .map((res: Response) => res.json())
-      .catch(this.handleError);
+      .catch(this.util.handleError);
   }
 
   handleStatusCode(status) {
@@ -60,32 +61,4 @@ export class ApplicationService {
     }
   }
 
-  private handleError(error: Response | any) {
-    let errors: any = [];
-    if (error instanceof Response) {
-      let body = error.json() || '';
-      errors = body.errors;
-      if (error.status) {
-        switch (error.status) {
-          case 400:
-            body = error.json() || '';
-            errors = body.errors;
-            break;
-          case 401:
-            errors = [{ message: 'Please log in.' }];
-            break;
-          case 403:
-            errors = [{ message: 'Access denied.' }];
-            break;
-          case 404:
-            errors = [{ message: 'The requested application is not found.' }];
-            break;
-          default:
-            errors = [];
-        }
-      }
-    }
-
-    return Observable.throw(errors);
-  }
 }
